00055: CROCK WORKAROUND FOR HANG AFTER COPYRIGHT BUG
AUTHOR:		Frank MacLachlan (fpm@crash.cts.com)

	On some systems, the boot process hangs after the copyright notice
is displayed.  This will "fix" the problem by placing two large delays in
the admittedly broken keyboard reset code (the comment for the "pcprobe"
and "pcattach" routines is "these are both bad jokes").

NOTE:
	This is just a workaround, and it isn't a crock because of any
fault on Frank's part... the probe and attach routines need to be rewritten.
One (anecdotal and without posted patches) fix to this problem that has been
suggested with replacing them with the other boot codes keyboard reset code
that performs the "Gate A20 hack".  This is supposed to fix the problem;
without patches or a way to repeat the problem, I'm stuck with it.