00047: FFLUSH STDIO LIBRARY FIX
AUTHOR:		Dave Rivers (rivers@ponds.uucp)

	I've discovered the following discrepency in the stdio library
w/ regards to opening a stream using  fopen("file","a+")  and later
calling fflush() on that stream. The fopen() call resets the flags to
__SRW because of the '+' in the mode string.  Later on, fflush() checks
to ensure that the stream being flushed is actually writable.  To do
this, it looks for the __SWR value in the flags, and *not* __SRW.  There
is a comment in stdio.h that implies that __SWR and __SRW will never be
set at the same time.