00032: KERNEL NAMELIST DATABASE CREATION/USE
AUTHOR:		Paul Kranenburg, David Greenman

	   There are several bugs in the kernel namelist db creation/use and
bugs in kvm.c which causes ps and several other utils to fail to work
correctly.

The kvm.c patches are based in part on, but supersede, the patches posted
on usenet by Paul Kranenburg [which introduced bugs of its own].
In order for kvm_mkdb to work correctly, vers.o MUST be the first object
linked in the kernel. The Makefile.i386 should read:
SYSTEM_LD= @${LD} -z -T FE000000 -o $@ -X vers.o ${SYSTEM_OBJS}
(or vers.o should be first in SYSTEM_OBJS)

And, of course all things that depend on libutil.a will have to be rebuilt.