00055: VARIOUS TTY DRIVER FIXES
AUTHOR:		Terry Lee (terry@uivlsi.csl.uiuc.edu)

	The backspace key returns DEL in non-X mode but returns ^H in X
mode.  According to the termcap entry, it should return ^H.  It also
makes more sense to have the backspace key return ^H and the delete key
return DEL.  SHIFT-TAB returns DEL.  This is not consistent with any
other keyboard/workstation I know of.  I believe it should still return
TAB and let the user-level program decide what to do with the SHIFT.  In
X mode (pc_xmode), the LEDs are no longer updated.