00500: CLONES OF WESTERN DIGITAL ETHERNET CARDS FAIL CHECKSUM
AUTHOR:		Mike Durkin (mdurkin@tsoft.sf-bay.org)

Danpex EW-2016 is an inexpensive (US$85.) allegedly WD8013-compatible
ethernet card.

I have one of these cards also and after fiddling a bit with the source
found that the checksum in the probe fails.  I commented it out and am
using the card successfully, NFS mounting to and from my Sun3s with no
problems.

This patch reports the failure, but doesn't fail the probe because of it.

NOTE:   Add "option WDCOMPAT" to enable this.  Do NOT do this if you will
        not be using a WD80x3 compatable board, since this will result in
        the kernel finding a board which isn't there.