00083: GATE A20 FIX FOR HANG AFTER COPYRIGHT & CROCK WORKAROUND FOR HANG AFTER COPYRIGHT BUG
AUTHOR:		Scotty ? & Frank MacLachlan (fpm@crash.cts.com)

	I combined these two patches into one, since they essentially
fix the same problem.  By default, the A20 fix is enabled, which fixes
most of the keyboard hanging problems.  However, if it does not work,
you need to define KB_HACK in your config file to enable to delay
routines in the keyboard reset code.

A20 INFO:
	The pc_probe code isn't reseting the a20. That is only done in
the bootstrap. The probe code is just trying to reset the keyboard
controller. Alas the original code uses general purpose routines for
issuing the commands. These general purpose routines read the keyboard
controller and return the value.

KB_HACK INFO:
	On some systems, the boot process hangs after the copyright notice
is displayed.  This will "fix" the problem by placing two large delays in
the admittedly broken keyboard reset code (the comment for the "pcprobe"
and "pcattach" routines is "these are both bad jokes").

NOTE:
	This is just a workaround, and it isn't a crock because of any
fault on Frank's part... the probe and attach routines need to be rewritten.