00045: FIXES SILO OVERFLOWS IF A 16550 IS INSTALLED
AUTHOR:		Rodney W. Grimes (rgrimes@agora.rain.com)

Here is a patch to fix silo overflows if you have 16550 uarts.  The
problem is due to the FIFO trigger level being set to 14 bytes.  This
only leaves 2 bytes after the interupt before it overflows.  This is
not enought time at high speeds.  By changing the trigger level to
4 bytes I have had a 38400 baud line running full speed with out any
silo overflows.