00040: CD-ROM FILES SEEM TO DISAPPAER
AUTHOR:		Scott Burris (scott@pita.cns.ucla.edu)

	The infamous "off by 1" problem hit two places in the isofs code.
There is a check to see that a directory entry doesn't span more than one
physical block -- if so, that's an error.  Well if the directory entry just
happened to end at the last byte of the block, the kernel throught it
crossed into the next block and bailed out assuming the CD-ROM wasn't following
the standard.  The effect is that not only can't you access the file associated
with this directory entry, you also can't see any files or directories located
after that entry.

This is relatively rare, because a series of directory entries must fit
perfectly into a 2048 byte CD-ROM block for this to trigger.