00063: KERNEL SPEEDUP #1
AUTHOR:		Poul-Henning Kamp

	This patch cuts a quite large chunk of the processing in the
kernel. For some unfortunate reasons pmap.c is a focal point from a
performance point of view, this patch improves things a bit, but it is
still the most used routine in the kernel.  The 16.6% is actually the
percentage 'system-time' is cut when doing a make kernel on my machine.