00065: SMALL BUFFER LEAK IN PHYSIO()
AUTHOR:		Mark Tinguely (tinguely@plains.nodak.edu)

	 There appears to be a small buffer leak in kern__physio.c. The
function physio() allocates a buf only for the lifetime of that
function. If the function exits on access error, the buf is not
released. This error was found while looking for M_TEMP allocation (ie.
may not be the direct cause of a known problem). This is a visible
fix. physio() is not called that often to cause a major memory loss.