00086: MAX_DOUBLE AND HUGE_VAL ARE TOO HUGE. (SIGNAL 6 ABORTS)
AUTHOR:		Nick Handel (nhandel@NeoSoft.com)
AUTHOR:		Peter da Silva (peter@NeoSoft.com)
AUTHOR:		Daniel Poirot (poirot@aio.jsc.nasa.gov)

Peter da Silva & Nick Handel:
	The MAX_DOUBLE and HUGE_VAL values are too big, big enough in
HUGE_VAL's case that you automatically get an abort signal on any attempt
to compare it to another value, resulting in "trap 6" messages and the like.

I heard something about bugs in atof and the like, so these are probably
not the final word, but they get a lot more stuff to compile and run.

[ Editor's note: ]

This is definately not the final word.  This is an interm fix until
the real fix can be released.  ALSO BE SURE TO READ THE TOP LEVEL
README FILE FOR THE PATCH KIT.  There are important instructions
on how to properly build things after installing patches that if
you do not follow this patch will not do anything for you.