10000: INCREASED tcp_sendpace AND tcp_recvspace TO 16K BYTES
AUTHOR:		David Greenman (davidg@implode.rain.com)

This is a experimental patch, it should increase the performance of
TCP/IP, but may actually slow it down when talking to certain
implementations of TCP/IP.

PLEASE report any problems with this patch to rgrimes@agora.rain.com,
as we need feedback.  If it seems to really help, we would be interested
in that too.