00101: FIXES /usr/share/man/makewhatis.sed IN BINDIST
AUTHOR:		Andy Valencia (ajv@csd.mot.com)
AUTHOR:		Rodney W. Grimes (rgrimes@agora.rain.com)

Rodney Grimes:
	This is a patch based upon Andy Valencia's patch to
/usr/othersrc/share/man/makewhatis.sed.  It REALLY fixes man -k.

This has been done in such a way that you do not need the ETCDIST installed
to use it.

It also gets us set up for the missing pieces of the bsd manual page sources
that need to be revised.  It does this by creating /usr/src/usr.share/...
And setting up the top level make file so that a top level make rebuilds
that makewhatis.sed and COPYRIGHT in /usr/share/man.

Still need to port the other man pages and clean them up, but this is
a start.  Also several of the man pages do not work properly with
man -k.  These are usually gnu man pages.