00070: MULTIPLE PROCESSES ON ONE PTY
AUTHOR:		Julian Elischer (julian@ref.tfs.com)

	Two processes each write to the same pty in a special case, each
sleeps waiting on theoutput queue.  Unfortunatly, the pty code sleeps
and wakes up on this same address because the sub-field it seeks on is
the first in the output queue structure, and thus has the same address
as the queue as a whole. The solution is to make the pty code sleep and
wakeup on another sub-field of the structure.  The result of the
confusion is that the two processes writing to the pty simply keep
waking each other up (at high priority, and then going to sleep
themselves at high priority)  As a result, all other work on the
processor comes to a stop.