00067: HANGING CONSOLE FIX
AUTHOR:		Holger Veit (veit@du9ds3.uni-duisburg.de)

	The problem is in /sys/kern/tty_pty. Ptsclose calls ttyclose
which invalidates a constty pointing to that tty slot correctly.
Ptcclose does not. Xterm -C grabs the console from elsewhere and assigns
it to its controlling pty. After end of xterm constty still points to
the now invalid struct tty. Ptcclose resets TS_CARR_ON, so that the
process that inherits this invalid console slot, goes to sleep, waiting
on ttopen, because there is "no carrier".