00117: Adds vmstat/iostat kernel support and more verbose device probe infoFix config so that you can have devices without irqs
AUTHOR:	Rodney W. Grimes (rgrimes@agora.rain.com) (vmstat/iostat, probe)
AUTHOR:	Phil Sutherland (philsuth@mycroft.dialix.oz.au) (config no irqs)
AUTHOR: (???@sun-lamp) (autoconf.c add support for floppy swap)

	This patch adds the symbol names to icu.s that vmstat expects
the interrupt counters to be called.  It also adds code to config
so that the names of the interrupts are written at the end of vectors.s
so vmstat can report real device names.  It also cleans up and enables
the logging of stray interrupts.  The counters for false interrupts
are added but the fix for them is not (the fix I have is not done
very good.) A false interrupt is when a device asserts it's interrupt
signal, then removes it before the 8259 can latch it.  This is the number
one cause of stray IRQ7's and IRQ15's.

	Additional device probe information is now printed.  This includes
ending I/O address (many drivers do not return the correct value from a
probe this still needs to be fixed), memory address and size, driver
flags passed in by config.

	It is now possible to configure devices that do not use an irq,
this is especially handy for the new lpa driver.

	During a boot from floppy the system now asks for a root floppy
disk.  Also now reports root on sd0a when new SCSI code is used.

NOTE:	The source for vmstat and iostat are added by other patches.