00135: Major new device support patch, if_ix, com3, com4, speaker, dcfclock
AUTHOR:	Rodney W. Grimes (rgrimes@agora.rain.com)

	Major rework of GENERICISA, converted to Julians SCSI code, added
devices com3, com4, lpt0, lpa0, lpa1 and pseudo-device speaker.

	Major rework of /dev/MAKEDEV, fixed up chown and chmod commands,
added all as an option to build all the most common devices.  Added new
device support.

	Fix the error from patch00097 with respect to NLPT should have been
NLPA in the second occurance in conf.c.  Added speaker support.

NOTE:  We are patching an installed file (/dev/MAKEDEV), this will need to
be propogated to /usr/src/etc when that patch is done.