00152: INCREASED tcp_sendpace AND tcp_recvspace TO 16K BYTES
AUTHOR:		David Greenman (davidg@implode.rain.com)
AUTHOR:		Rodney W. Grimes (rgrimes@agora.rain.com)

This is a experimental patch, it should increase the performance of
TCP/IP, but may actually slow it down when talking to certain
implementations of TCP/IP.

PLEASE report any problems with this patch to rgrimes@agora.rain.com,
as we need feedback.  If it seems to really help, we would be interested
in that too.

This patch has been revised, most people should now install this patch,
if it creates a problem for you please let us know.  You can work
around this by adding an options "TCP_SMALLSPACE" to you kernel config
file.