00158: Complete rewrite of the interrupt and spl code (speed improvement)
AUTHOR:	Bruce Evans (bde@runx.oz.au)
AUTHOR:	Rodney W. Grimes (rgrimes@agora.rain.com) (intergrate, and cleanup)

	Way more than what could be put here.  This is a complete rewrite
of the interrupt handling code and the spl mechanisms.  This code is much
faster and should completly eliminate silo over flows for speeds up to
19200 on even the slowest of machines.  The only other part needed to
completly eliminate the silo over flows is a new com driver that is
provided in a later patch.

	I have reworked some of Bruce Evans code to enhance certain areas
that were not functional in the code provided by Bruce, which include the
ability to build and run the GENERICISA kernel.
							Rod Grimes