00141: There are *two* bugs in the 386BSD fifo code
AUTHOR:	Jay Fenlason <hack@datacube.com>

1: the fi_readers and fi_writers fields of the fifoinfo structure were not being
initialized to 0.  This caused the driver to not sleep the first process to
open the fifo--it thought there was already another process to talk to
(most of the time.)

2: fifo_open() was calling tsleep() without unlocking the inode of the fifo
file.  This caused *any* subsequent access to the file (even an ls (!)) to
hang forever.  Note that this bug was usually masked by bug #1 above.

Note that this patch is suspect:  I think this may result in both the
fifo-reader and the fifo-writer locking the inode at the same time (a no-no).
The kernel might panic if -DDIAGNOSTIC was used (I haven't checked).  I don't
know kernel internals well enough to tell.

Also note that the structure fifo_vnodeops is completely unused (replaced by
the structure fifo_inodeops in ufs/ufs_vnops.c).  The same applies to the
functions fifo_lock() and fifo_unlock().

Date: Wed, 31 Mar 93 17:06:01 EST
Index: sys.386bsd/kern 386BSD-0.1